+++ demisto_sdk/commands/content_graph/common.py
@@ -25,6 +25,7 @@ UNIFIED_FILES_SUFFIXES = [".yml", ".json"]
 
 
 class Neo4jRelationshipResult(NamedTuple):
+    node_from: graph.Node
     relationships: List[graph.Relationship]
     nodes_to: List[graph.Node]
 
@@ -77,6 +78,7 @@ class ContentType(str, enum.Enum):
     XSIAM_REPORT = "XSIAMReport"
     WIZARD = "Wizard"
     XDRC_TEMPLATE = "XDRCTemplate"
+    LAYOUT_RULE = "LayoutRule"
 
     @property
     def labels(self) -> List[str]: