+++ demisto_sdk/commands/content_graph/interface/neo4j/queries/nodes.py
@@ -4,19 +4,21 @@ from typing import Any, Dict, Iterable, List, Optional
 from neo4j import Transaction, graph
 
 from demisto_sdk.commands.common.constants import MarketplaceVersions
-from demisto_sdk.commands.content_graph.common import SERVER_CONTENT_ITEMS, ContentType
+from demisto_sdk.commands.content_graph.common import (
+    SERVER_CONTENT_ITEMS,
+    ContentType,
+    RelationshipType,
+)
 from demisto_sdk.commands.content_graph.interface.neo4j.queries.common import (
-    intersects,
     run_query,
     to_neo4j_map,
-    versioned,
 )
 
 logger = logging.getLogger("demisto-sdk")
 
 NESTING_LEVEL = 5
 
-CREATE_CONTENT_ITEM_NODES_BY_TYPE_TEMPLATE = """
+CREATE_CONTENT_ITEM_NODES_BY_TYPE_TEMPLATE = """// Creates/overrides existing content items with labels {labels}
 UNWIND $data AS node_data
 MERGE (n:{labels}{{
     object_id: node_data.object_id,
@@ -28,35 +30,18 @@ SET n = node_data,  // override existing data
 WITH n
     OPTIONAL MATCH (n)-[r]->()
     DELETE r
-RETURN count(n) AS nodes_created
-"""
+RETURN count(n) AS nodes_created"""
 
 
-CREATE_NODES_BY_TYPE_TEMPLATE = """
+CREATE_NODES_BY_TYPE_TEMPLATE = """// Creates/overrides existing nodes with labels {labels}
 UNWIND $data AS node_data
 MERGE (n:{labels}{{object_id: node_data.object_id}})
 SET n = node_data,  // override existing data
     n.not_in_repository = false
-RETURN count(n) AS nodes_created
-"""
+RETURN count(n) AS nodes_created"""
 
 
-FIND_DUPLICATES = f"""
-MATCH (a:{ContentType.BASE_CONTENT})
-MATCH (b:{ContentType.BASE_CONTENT}{'{node_id: a.node_id}'})
-WHERE
-    id(a) <> id(b)
-AND
-    {intersects('a.marketplaces', 'b.marketplaces')}
-AND
-    {versioned('a.toversion')} >= {versioned('b.fromversion')}
-AND
-    {versioned('b.toversion')} >= {versioned('a.fromversion')}
-RETURN count(b) > 0 AS found_duplicates
-"""
-
-
-REMOVE_SERVER_NODES_BY_TYPE = """
+REMOVE_SERVER_NODES_BY_TYPE = """// Removes parsed server nodes of type {content_type} (according to constants)
 MATCH (a)
 WHERE (a:{label} OR a.content_type = "{content_type}")
 AND a.not_in_repository = true
@@ -64,19 +49,80 @@ AND any(
     identifier IN [a.object_id, a.name]
     WHERE toLower(identifier) IN {server_content_items}
 )
-DETACH DELETE a
-"""
+DETACH DELETE a"""
 
 
-REMOVE_EMPTY_PROPERTIES = """CALL apoc.periodic.iterate(
+REMOVE_EMPTY_PROPERTIES = """// Removes string properties with empty values ("") from nodes
+CALL apoc.periodic.iterate(
     "MATCH (n) RETURN n",
     "WITH n, [key in keys(n) WHERE n[key] = '' | [key, null]] as nullifiers
     WHERE size(nullifiers) <> 0
     WITH n, apoc.map.fromPairs(nullifiers) as nullifyMap
     SET n += nullifyMap",
     {batchSize:30000, parallel:true, iterateList:true}
-);
+);"""
+
+
+def get_relationships_to_preserve(
+    tx: Transaction,
+    pack_ids: List[str],
+) -> List[Dict[str, Any]]:
+    query = f"""// Gets the relationships to preserve before removing packs
+MATCH (s)-[r]->(t)-[:{RelationshipType.IN_PACK}]->(p)
+WHERE NOT (s)-[:{RelationshipType.IN_PACK}]->(p)
+AND p.object_id in {pack_ids}
+RETURN id(s) as source_id, type(r) as r_type, properties(r) as r_properties, t as target
+
+UNION
+
+MATCH (s)-[r]->(t)<-[:{RelationshipType.HAS_COMMAND}]-()-[:{RelationshipType.IN_PACK}]->(p)
+WHERE NOT (s)-[:{RelationshipType.IN_PACK}]->(p)
+AND p.object_id in {pack_ids}
+RETURN id(s) as source_id, type(r) as r_type, properties(r) as r_properties, t as target
+
+UNION
+
+MATCH (s)-[r]->(t)
+WHERE NOT (s)-[:{RelationshipType.IN_PACK}]->(t)
+AND t.object_id in {pack_ids}
+RETURN id(s) as source_id, type(r) as r_type, properties(r) as r_properties, t as target"""
+    return run_query(tx, query).data()
+
+
+def remove_packs_before_creation(
+    tx: Transaction,
+    pack_ids: List[str],
+) -> None:
+    query = f"""// Removes packs commands before recreating them
+MATCH (c)<-[:{RelationshipType.HAS_COMMAND}]-()-[:{RelationshipType.IN_PACK}]->(p)
+WHERE p.object_id IN {pack_ids}
+OPTIONAL MATCH (c)<-[:{RelationshipType.HAS_COMMAND}]-()-[:{RelationshipType.IN_PACK}]->(p2)
+WHERE NOT p2.object_id IN {pack_ids}
+WITH c, p2
+WHERE p2 IS NULL
+DETACH DELETE c
 """
+    run_query(tx, query)
+    query = f"""// Removes packs and their content items before recreating them
+MATCH (n)-[:{RelationshipType.IN_PACK}]->(p)
+WHERE p.object_id in {pack_ids}
+DETACH DELETE n, p"""
+    run_query(tx, query)
+
+
+def return_preserved_relationships(
+    tx: Transaction, rels_to_preserve: List[Dict[str, Any]]
+) -> None:
+    query = """// Returns the preserved relationships
+UNWIND $rels_data AS rel_data
+MATCH (s) WHERE id(s) = rel_data.source_id
+OPTIONAL MATCH (t{object_id: rel_data.target.object_id, content_type: rel_data.target.content_type})
+WITH s, t, rel_data
+WHERE NOT t IS NULL
+CALL apoc.create.relationship(s, rel_data.r_type, rel_data.r_properties, t)
+YIELD rel
+RETURN rel"""
+    run_query(tx, query, rels_data=rels_to_preserve)
 
 
 def create_nodes(
@@ -102,11 +148,6 @@ def remove_server_nodes(tx: Transaction) -> None:
         run_query(tx, query)
 
 
-def duplicates_exist(tx) -> bool:
-    result = run_query(tx, FIND_DUPLICATES).single()
-    return result["found_duplicates"]
-
-
 def create_nodes_by_type(
     tx: Transaction,
     content_type: ContentType,
@@ -119,7 +160,7 @@ def create_nodes_by_type(
         query = CREATE_NODES_BY_TYPE_TEMPLATE.format(labels=labels)
     result = run_query(tx, query, data=data).single()
     nodes_count: int = result["nodes_created"]
-    logger.info(f"Created {nodes_count} nodes of type {content_type}.")
+    logger.debug(f"Created {nodes_count} nodes of type {content_type}.")
 
 
 def _match(
@@ -152,11 +193,10 @@ def _match(
             where.append("AND")
         if marketplace:
             where.append(f"'{marketplace}' IN node.marketplaces")
-    query = f"""
-    MATCH (node{content_type_str}{params_str})
+    query = f"""// Retrieves nodes according to given parameters.
+MATCH (node{content_type_str}{params_str})
     {" ".join(where)}
-    RETURN node
-    """
+RETURN node"""
     if ids_list:
         query = "UNWIND $filter_list AS node_id\n" + query
 
@@ -169,10 +209,9 @@ def _match(
 
 
 def delete_all_graph_nodes(tx: Transaction) -> None:
-    query = """
-    MATCH (n)
-    DETACH DELETE n
-    """
+    query = """// Deletes all graph nodes and relationships
+MATCH (n)
+DETACH DELETE n"""
     run_query(tx, query)