+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
+
 ## Unreleased
+* **Breaking change**: Removed usage of `pipenv`, `isort` and `autopep8` in the **split** and **download** commands. Removed the `--no-pipenv` flag. Please see https://xsoar.pan.dev/docs/tutorials/tut-setup-dev-remote for the recommended environment setup.
+* Fixed an issue in **prepare-content** command where large code lines were broken.
+* Fixed an issue where git-*renamed_files* were not retrieved properly.
+* Fixed an issue where test dependencies were calculated in all level dependencies calculation.
+* Added formatting and validation to XSIAM content types.
+* Fixed an issue where several XSIAM content types were not validated when passing the `-a` flag.
+* Added a UUID to name mapper for **download** it replaces UUIDs with names on all downloaded files.
+* Updated the demisto-py to v3.2.6 which now supports basic proxy authentication.
+* Improved the message shown when using **upload** and overwriting packs.
+* Added support for the **Layout Rule** content type in the id-set and the content graph.
+* Updated the default general `fromVersion` value on **format** to `6.8.0`
+* Fixed an issue where **lint** sometimes failed when using the `-cdam` flag due to wrong file duplications filtering.
+* Updated **validate** to use the graph when running with `--graph` flag.
 
 
 ## 1.9.0